+++ pkg/analysis_server/lib/src/services/correction/dart/change_to.dart
@@ -136,7 +136,7 @@ class ChangeTo extends ResolvedCorrectionProducer {
       }
       // Check elements from imports.
       for (var importElement
-          in unitResult.libraryElement2.firstFragment.libraryImports2) {
+          in unitResult.libraryElement2.firstFragment.libraryImports) {
         if (importElement.prefix2?.element.name == prefixName) {
           var namespace = getImportNamespace(importElement);
           finder._updateList(namespace.values);
@@ -243,7 +243,7 @@ class ChangeTo extends ResolvedCorrectionProducer {
       }
       // Check unprefixed imports.
       for (var importElement
-          in unitResult.libraryElement2.firstFragment.libraryImports2) {
+          in unitResult.libraryElement2.firstFragment.libraryImports) {
         if (importElement.prefix2?.element.name == prefixName) {
           var namespace = getImportNamespace(importElement);
           finder._updateList(namespace.values);