+++ pkg/analysis_server/lib/src/services/correction/namespace.dart
@@ -44,7 +44,7 @@ LibraryImport? _getImportElement(
   var usedLibrary = element.library;
   // find ImportElement that imports used library with used prefix
   List<LibraryImport>? candidates;
-  for (var libraryImport in libraryElement.firstFragment.libraryImports2) {
+  for (var libraryImport in libraryElement.firstFragment.libraryImports) {
     // required library
     if (libraryImport.importedLibrary != usedLibrary) {
       continue;